requirements.txt
chromadb==0.5.5
duckduckgo_search==6.2.6
fastapi==0.112.0
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
langchain==0.2.12
langchain-community==0.2.11
langchain-google-genai==1.0.8
langchainhub==0.1.20
python==3.11.9
pdfplumber==0.11.3
psycopg2==2.9.9
PyPDF2==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
streamlit==1.37.1
tiktoken==0.7.0
transformers==4.44.0
unstructured==0.15.3
uvicorn==0.30.5
wikipedia==1.4.0